requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.4
requests~=2.28.1
coreapi~=2.3.3
django-allauth~=0.52.0
Django~=3.2.16
PyYAML~=6.0
pika~=1.3.1